Create better test Prometheus metric data
This creates metrics of the form

    test{instance="test-metric-XXX", job="testdata"}

Where "XXX" is a zero based index that increments from 0 to the number
of time series we are asked to create.  This gives metrics that look a
little more like real data.

External Thanos labels added are {replica="42", monitor="test"}.

Metric values are now nanoseconds since the epoch.