Fix chunk metadata's maxTime
Set chunk's max time to the timepoint of the maximum element in the
chunk (i.e. start + length). The previous behavior was adding ampleInterval
which is equivalent to just a single datapoint added. Suprisingly it
kind of works with the buggy version in  prometheus's graph view but has weird
behavior in the table view. E.g. when generating data in 1s intervals it
is able to show data from 0s to +5m10s (31 data points) but cannot show
any data thereafter.